% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData: Import views directly from HTLN bladderpod database}
\usage{
importData(
  instance = c("local", "server"),
  server = NA,
  new_env = TRUE,
  path = NA,
  name = "MoBlad_BloodyHill"
)
}
\arguments{
\item{instance}{Specify whether you are connecting to the local instance or
server.
\describe{
\item{"local"}{Default. Connects to local install of frontend database}
\item{"server"}{Connects to main backend on server. Note that you must
have permission to access the server, and connection speeds are likely to be
much slower than the local instance. You must also be connected to VPN or
NPS network.}}}

\item{server}{Quoted name of the server to connect to, if instance = "server".
Valid input is the server address to connect to the main database. If
connecting to the local instance, leave blank.}

\item{new_env}{Logical. Specifies which environment to store views in.
If \code{TRUE}(Default), stores views in VIEWS_HTLN_BLDP environment.
If \code{FALSE}, stores views in global environment}

\item{path}{Path to where the frontend database, if instance = "local".}

\item{name}{Character. Specifies the name of the database.}
}
\value{
HTLN bladderpod database views in specified environment
}
\description{
This function imports all views in the HTLN MoBlad_BloodyHill
Microsoft Access DB. Each view is added to a VIEWS_HTLN_BLDP environment
in your workspace, or to your global environment based on whether
new_env = TRUE or FALSE.
}
\examples{
\dontrun{
# Import using default settings of local instance, server = 'localhost' and add VIEWS_MIDN_NCBN environment
importData()

# Import using computer name (# should be real numbers)
importData(server = "HTLN-######", new_env = TRUE)

# Import from main database on server
importData(server = "INP###########\\\\########", instance = "server", new_env = TRUE)
}

}
